+++ drivers/staging/iio/accel/lis3l02dq_core.c
@@ -533,7 +533,8 @@ static struct iio_chan_spec lis3l02dq_channels[] = {
 	IIO_CHAN_SOFT_TIMESTAMP(3)
 };
 
-static ssize_t lis3l02dq_read_event_config(struct iio_dev *indio_dev,
+
+static int lis3l02dq_read_event_config(struct iio_dev *indio_dev,
 					   u64 event_code)
 {
